+++ Models/Transactions.cs
@@ -9,7 +9,7 @@ namespace PersonalFinance.Models
         public string TrsCode { get; set; }
         public string TrsTitle { get; set; }
         public DateTime TrsDateTime { get; set; }
-        public float TrsValue { get; set; }
+        public double TrsValue { get; set; }
         public string TrsNote { get; set; }
     }
 }
